+++ packages/website-generator/src/index.ts
@@ -1,5 +1,7 @@
 import { generate } from './generate';
+import { serve } from './serve';
 
 export default {
-  generate
+  generate,
+  serve,
 };